feat: WaitingScreen 模块初始化等待界面 + TimeManager 定时RTC同步
- 新增 WaitingScreen - 4G 模块初始化期间显示英文等待界面
- 启动时 push WaitingScreen，完成后 replace 为主菜单
- TimeManager 增加定时重新同步（每6小时重新 AT+CCLK?）
- handleRecvURC 解析失败时加分段诊断日志
- available() 非CIP路径打印未知行供诊断
- publish() > 等待循环转发 +MSUB: 到接收队列

+++ pass_key/src/mqtt/MQTTAtDriver.h
@@ -28,11 +28,11 @@ class Air780epDriver;
 class MQTTAtDriver
 {
 public:
-    // +QMTRECV 入站消息
+    // +MSUB 入站消息结构（与 MQTTManager::PendingMsg 保持内存布局一致）
     struct RecvMessage {
         char topic[128];
         uint8_t payload[512];
-        uint16_t payloadLen;
+        unsigned int payloadLen;
     };
 
     MQTTAtDriver(Air780epDriver *driver);
